chore: improve interopability
- Enable TypeScript esModuleInterop

This enables a better integration with Babel and this follows import
rules of NodeJS.

We should only import the default entry from a cjs package:
```
import entry from 'cjs-module'
```

Please avoid named imports or `* as name` imports.

- Enable TypeScript isolatedModules, and add `import type` statments

This enables the support of TypeScript type-dropping for third-party
tools such as Babel.

This is going to be the standard way to compile a TS project.

- Emit ES modules with TSC

This enables webpack to reduce build size by 1ko.
This also improves compatibility with other bundlers.

- Ensure consistent filename casting

- Remove @types/uws