feat: do not require package.json for execution
- Do not require `package.json` in the source code

This was required to get the current version of the package.
We now use a compile-time constant to set the package version.

semantic-release relies on semantic-release/npm plugin.
This plugin executes `npm version` then `npm publish`.
We rely on the npm hook `postversion` to build the source once
`package.json` is up•-to-date.

See semantic-release FAQ:
https://semantic-release.gitbook.io/semantic-release/support/faq